backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic[email]
python-multipart==0.0.12
aiosmtplib==3.0.1
python-dotenv==1.0.1
Jinja2==3.1.4
